% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mumm.R
\name{mumm}
\alias{mumm}
\title{Fit Multiplicative Mixed Models with TMB}
\usage{
mumm(formula, data)
}
\arguments{
\item{formula}{a two-sided formula object describing the linear fixed-effects and random-effects part
together with the multiplicative part. The respons is on the left of a ~ operator and the terms which are
separated by + operators are on the right. The random-effect terms are recognized by vertical bars "|",
separating an expression for a model matrix and a grouping factor. The syntax for the multiplicative terms
is 'mp("random effect","fixed effect")'.}

\item{data}{a data frame containing the variables in the formula.}
}
\value{
An object of class mumm.
}
\description{
Fit a multiplicative mixed-effects model to data with use of the Template Model Builder.
}
\details{
Fit a multiplicative mixed model via maximum likelihood with use of the Template Model Builder.
A multiplicative mixed model is here considered as a model with a linear mixed model part and one or more
multiplicative terms. A multiplicative term is here defined as a product of a random effect and a fixed effect,
i.e. a term that models a part of the interaction as a random coefficient model based on linear regression
on the fixed effect.
}
\examples{
library(SensMixed)
fit = mumm(Colourbalance ~ 1 + Picture + (1|Assessor) + (1|Assessor:Picture) +
           mp(Assessor,Picture) ,data = TVbo)

}

