% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mumm.R
\name{mumm}
\alias{mumm}
\title{Fit Multiplicative Mixed Models with TMB}
\usage{
mumm(formula, data, report = FALSE)
}
\description{
Fit a multiplicative mixed-effects model to data with use of the Template Model Builder.
}

